Solve dropdown and text field bug
